refactor: reorganize shared/ folder by domain
Replace technical grouping (contexts/, hooks/ai/, components/) with
domain-based folders (ai/, speech/, language/, snackbar/, theme/).

- Merge AIProvider, AI hooks, AI capabilities, and AICapabilitiesList into shared/ai/
- Move SpeechProvider files into shared/speech/
- Move LanguageProvider + language utils into shared/language/
- Move SnackbarProvider into shared/snackbar/
- Move ThemeProvider into shared/theme/
- Keep generic components/, hooks/, utils/, and testing/ unchanged
- Update all import paths across app/, features/, pages/, and shared/
