+++ src/app/AppProviders.tsx
@@ -1,8 +1,8 @@
-import { AIProvider } from "@shared/contexts/AIProvider/AIProvider";
-import { LanguageProvider } from "@shared/contexts/LanguageProvider/LanguageProvider";
-import { SnackbarProvider } from "@shared/contexts/SnackbarProvider/SnackbarProvider";
-import { SpeechProvider } from "@shared/contexts/SpeechProvider/SpeechProvider";
-import { ThemeProvider } from "@shared/contexts/ThemeProvider/ThemeProvider";
+import { AIProvider } from "@shared/ai/AIProvider";
+import { LanguageProvider } from "@shared/language/LanguageProvider";
+import { SnackbarProvider } from "@shared/snackbar/SnackbarProvider";
+import { SpeechProvider } from "@shared/speech/SpeechProvider";
+import { ThemeProvider } from "@shared/theme/ThemeProvider";
 import type { ReactNode } from "react";
 
 export interface AppProvidersProps {
