+++ src/app/drawers/SettingsDrawer/panels/AISettings.tsx
@@ -1,9 +1,9 @@
 import Stack from "@mui/material/Stack";
 import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
-import { AICapabilitiesList } from "@shared/components/AICapabilitiesList";
-import { useAI } from "@shared/contexts/AIProvider/useAI";
-import { isRewriterSupported } from "@shared/hooks/ai/ai-capabilities";
+import { AICapabilitiesList } from "@shared/ai/AICapabilitiesList";
+import { useAI } from "@shared/ai/useAI";
+import { isRewriterSupported } from "@shared/ai/ai-capabilities";
 
 export function AISettings() {
   const { sharedContext, setSharedContext } = useAI();
