+++ src/features/board/hooks/useBoardTranslation.ts
@@ -1,5 +1,5 @@
-import { useLanguage } from "@shared/contexts/LanguageProvider/useLanguage";
-import { useTranslator } from "@shared/hooks/ai/useTranslator";
+import { useTranslator } from "@shared/ai/useTranslator";
+import { useLanguage } from "@shared/language/useLanguage";
 import { useEffect, useState } from "react";
 import { updateBoardStrings, withBoardsDB } from "../db/boards-db";
 import type { Board, BoardButton } from "../types";
