+++ src/features/board/hooks/useButtonActivation.ts
@@ -1,4 +1,4 @@
-import { useSpeech } from "@shared/contexts/SpeechProvider/useSpeech";
+import { useSpeech } from "@shared/speech/useSpeech";
 import { useAudio } from "@shared/hooks/useAudio";
 import type { BoardAction, BoardButton } from "../types";
 import type { UseBoardNavigationReturn } from "./useBoardNavigation";
