+++ src/shared/ai/AIContext.ts
