+++ src/shared/ai/AIProvider.tsx
