+++ src/shared/ai/ai-capabilities.ts
