+++ src/shared/ai/useAI.ts
